Move and extend inventory helper
